Implement enforced url params && adapt usages
